Renames config.js to constants.js
